Server_Flask/requirements.txt
# flask==2.3.3
# torch==2.0.1
# torchvision==0.15.2
# opencv-python==4.8.1.78
# pillow==10.0.0
# numpy==1.24.3
# requests==2.31.0
# ultralytics==8.0.186

flask
pillow
torch>=1.10.0
torchvision
numpy
ultralytics
